backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.1
langchain==0.0.335
langchain-community==0.0.10
pydantic==2.4.2
python-multipart==0.0.6
pypdf==3.17.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2